--- CHIP-8/CHIP-8/bindkeys.ui
@@ -1,195 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>BindKeys</class>
- <widget class="QDialog" name="BindKeys">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>424</width>
-    <height>362</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Change Key Binds</string>
-  </property>
-  <layout class="QGridLayout" name="gridLayout">
-   <item row="11" column="4">
-    <widget class="QPushButton" name="Close">
-     <property name="text">
-      <string>Close</string>
-     </property>
-    </widget>
-   </item>
-   <item row="10" column="1">
-    <widget class="QLabel" name="label_C">
-     <property name="text">
-      <string>C</string>
-     </property>
-    </widget>
-   </item>
-   <item row="4" column="1">
-    <widget class="QLabel" name="label_3">
-     <property name="text">
-      <string>3</string>
-     </property>
-    </widget>
-   </item>
-   <item row="2" column="3">
-    <widget class="QLabel" name="label_2">
-     <property name="text">
-      <string>2</string>
-     </property>
-    </widget>
-   </item>
-   <item row="11" column="2">
-    <widget class="QPushButton" name="RestoreDefaults">
-     <property name="text">
-      <string>Restore Defaults</string>
-     </property>
-    </widget>
-   </item>
-   <item row="2" column="1">
-    <widget class="QLabel" name="label_1">
-     <property name="text">
-      <string>1</string>
-     </property>
-    </widget>
-   </item>
-   <item row="7" column="3">
-    <widget class="QLabel" name="label_S">
-     <property name="text">
-      <string>S</string>
-     </property>
-    </widget>
-   </item>
-   <item row="7" column="1">
-    <widget class="QLabel" name="label_A">
-     <property name="text">
-      <string>A</string>
-     </property>
-    </widget>
-   </item>
-   <item row="9" column="1">
-    <widget class="QLabel" name="label_Z">
-     <property name="text">
-      <string>Z</string>
-     </property>
-    </widget>
-   </item>
-   <item row="5" column="3">
-    <widget class="QLabel" name="label_W">
-     <property name="text">
-      <string>W</string>
-     </property>
-    </widget>
-   </item>
-   <item row="10" column="3">
-    <widget class="QLabel" name="label_V">
-     <property name="text">
-      <string>V</string>
-     </property>
-    </widget>
-   </item>
-   <item row="8" column="1">
-    <widget class="QLabel" name="label_D">
-     <property name="text">
-      <string>D</string>
-     </property>
-    </widget>
-   </item>
-   <item row="4" column="3">
-    <widget class="QLabel" name="label_4">
-     <property name="text">
-      <string>4</string>
-     </property>
-    </widget>
-   </item>
-   <item row="8" column="3">
-    <widget class="QLabel" name="label_F">
-     <property name="text">
-      <string>F</string>
-     </property>
-    </widget>
-   </item>
-   <item row="6" column="1">
-    <widget class="QLabel" name="label_E">
-     <property name="text">
-      <string>E</string>
-     </property>
-    </widget>
-   </item>
-   <item row="9" column="3">
-    <widget class="QLabel" name="label_X">
-     <property name="text">
-      <string>X</string>
-     </property>
-    </widget>
-   </item>
-   <item row="5" column="1">
-    <widget class="QLabel" name="label_Q">
-     <property name="text">
-      <string>Q</string>
-     </property>
-    </widget>
-   </item>
-   <item row="6" column="3">
-    <widget class="QLabel" name="label_R">
-     <property name="text">
-      <string>R</string>
-     </property>
-    </widget>
-   </item>
-   <item row="2" column="2">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_1"/>
-   </item>
-   <item row="2" column="4">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_2"/>
-   </item>
-   <item row="4" column="2">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_3"/>
-   </item>
-   <item row="4" column="4">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_4"/>
-   </item>
-   <item row="5" column="2">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_Q"/>
-   </item>
-   <item row="5" column="4">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_W"/>
-   </item>
-   <item row="6" column="2">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_E"/>
-   </item>
-   <item row="6" column="4">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_R"/>
-   </item>
-   <item row="7" column="2">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_A"/>
-   </item>
-   <item row="7" column="4">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_S"/>
-   </item>
-   <item row="8" column="2">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_D"/>
-   </item>
-   <item row="8" column="4">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_F"/>
-   </item>
-   <item row="9" column="2">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_Z"/>
-   </item>
-   <item row="9" column="4">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_X"/>
-   </item>
-   <item row="10" column="2">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_C"/>
-   </item>
-   <item row="10" column="4">
-    <widget class="QKeySequenceEdit" name="keySequenceEdit_V"/>
-   </item>
-  </layout>
- </widget>
- <resources/>
- <connections/>
-</ui>
