--- CHIP-8/CHIP-8/keybinds.cpp
@@ -1,46 +1,0 @@
-#include "keybinds.h"
-
-KeyBinds::KeyBinds(QObject *parent)
-    : QAbstractItemModel(parent)
-{
-}
-
-QVariant KeyBinds::headerData(int section, Qt::Orientation orientation, int role) const
-{
-    // FIXME: Implement me!
-}
-
-QModelIndex KeyBinds::index(int row, int column, const QModelIndex &parent) const
-{
-    // FIXME: Implement me!
-}
-
-QModelIndex KeyBinds::parent(const QModelIndex &index) const
-{
-    // FIXME: Implement me!
-}
-
-int KeyBinds::rowCount(const QModelIndex &parent) const
-{
-    if (!parent.isValid())
-        return 0;
-
-    // FIXME: Implement me!
-}
-
-int KeyBinds::columnCount(const QModelIndex &parent) const
-{
-    if (!parent.isValid())
-        return 0;
-
-    // FIXME: Implement me!
-}
-
-QVariant KeyBinds::data(const QModelIndex &index, int role) const
-{
-    if (!index.isValid())
-        return QVariant();
-
-    // FIXME: Implement me!
-    return QVariant();
-}
