+++ Chip8Redo/keybinds.h
@@ -1,0 +1,21 @@
+#ifndef KEYBINDS_H
+#define KEYBINDS_H
+
+#include <QAbstractItemModel>
+
+class KeyBinds : public QAbstractItemModel
+{
+    Q_OBJECT
+
+public:
+    KeyBinds(QObject *parent = nullptr);
+
+    QVariant headerData(int section, Qt::Orientation orientation, int role = Qt::DisplayRole) const override;
+    QModelIndex index(int row, int column, const QModelIndex &parent = QModelIndex()) const override;
+    QModelIndex parent(const QModelIndex &index) const override;
+    int rowCount(const QModelIndex &parent = QModelIndex()) const override;
+    int columnCount(const QModelIndex &parent = QModelIndex()) const override;
+    QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const override;
+};
+
+#endif // KEYBINDS_H
